remove statsd_host declaration for Zuul
Via the standard search-and-replace, we were referring to a
'graphite.elasticdb.org' as a statsd server.
For some reason, exact specifics unknown, this had the affect of
causing the Zuul service to fail to start as the user zuul.

The root user could run Zuul - but zuul user would fail to create
gearman server.
With this change, the zuul user can run gearman and zuul.

Change-Id: Id2bb8bceb84c73951e1f4dd48c9fc8a20aa837b2