update virtualenv to at least 13.1.0
Virtualenv contains its own version of pip. The virtualenv version
dictate the pip version in use inside venvs.

Our virtualenv was stale, because our system-config is old.

The puppet logic seems to ensure that the virtualenv is at least
the version specified.
This seems to affect centos6 where the virtualenv version is much
older than in trusty images.

Change-Id: Ie608276ae477750a7d06a1855a55d5bc49fa1e35